Move resizable-column to directives folder
